requirements.txt
httplib2
google-api-python-client
GoogleAppEngineCloudStorageClient
